Fixes C++ Code

+++ libs/ui/GraphicBufferAllocator.cpp
@@ -65,7 +65,7 @@ void GraphicBufferAllocator::dump(String8& result) const
         const alloc_rec_t& rec(list.valueAt(i));
         if (rec.size) {
             snprintf(buffer, SIZE, "%10p: %7.2f KiB | %4u (%4u) x %4u | %8X | 0x%08x\n",
-                    list.keyAt(i), rec.size/1024.0f,
+                    list.keyAt(i), static_cast<double>(rec.size/1024.0f),
                     rec.width, rec.stride, rec.height, rec.format, rec.usage);
         } else {
             snprintf(buffer, SIZE, "%10p: unknown     | %4u (%4u) x %4u | %8X | 0x%08x\n",
@@ -75,7 +75,7 @@ void GraphicBufferAllocator::dump(String8& result) const
         result.append(buffer);
         total += rec.size;
     }
-    snprintf(buffer, SIZE, "Total allocated (estimate): %.2f KB\n", total/1024.0f);
+    snprintf(buffer, SIZE, "Total allocated (estimate): %.2f KB\n", static_cast<double>(total/1024.0f));
     result.append(buffer);
     if (mAllocDev->common.version >= 1 && mAllocDev->dump) {
         mAllocDev->dump(mAllocDev, buffer, SIZE);
